requirements.txt
flask
python-pptx
PyPDF2
python-docx
werkzeug
